pyproject.toml
[project]
name = "mcp-discord-auto-response"
version = "0.1.0"
description = "Discord Auto-Response MCP Server - Automatically respond to friend messages"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "discord.py>=2.0.0",
    "certifi>=2025.11.12",
]

[project.scripts]
mcp-server = "template_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/template_server"]

[dependency-groups]
dev = []
